refactor(#130): 기존의 계획 상세보기가 위치하던 곳에서 MypageController로 모두 이동
