--- src/main/java/com/sideproject/hororok/member/application/MemberService.java
@@ -1,47 +1,0 @@
-package com.sideproject.hororok.member.application;
-
-import com.sideproject.hororok.cafe.dto.CafeDto;
-import com.sideproject.hororok.keword.domain.Keyword;
-import com.sideproject.hororok.keword.domain.enums.Category;
-import com.sideproject.hororok.keword.dto.CategoryKeywordsDto;
-import com.sideproject.hororok.member.dto.response.MyPagePlanDetailResponse;
-import com.sideproject.hororok.plan.application.PlanCafeService;
-import com.sideproject.hororok.plan.application.PlanKeywordService;
-import com.sideproject.hororok.plan.domain.Plan;
-import com.sideproject.hororok.plan.domain.enums.MatchType;
-import com.sideproject.hororok.plan.domain.enums.PlanCafeMatchType;
-import com.sideproject.hororok.plan.domain.repository.PlanKeywordRepository;
-import com.sideproject.hororok.plan.domain.repository.PlanRepository;
-import com.sideproject.hororok.plan.dto.PlanDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class MemberService {
-
-    private final PlanRepository planRepository;
-
-    private final PlanKeywordService planKeywordService;
-    private final PlanCafeService planCafeService;
-
-    public MyPagePlanDetailResponse planDetail(Long planId) {
-
-        Plan findPlan = planRepository.getById(planId);
-        List<Keyword> findKeywords = planKeywordService.getKeywordsByPlanId(planId);
-        CategoryKeywordsDto categoryKeywords = new CategoryKeywordsDto(findKeywords);
-        List<CafeDto> findSimilarCafes = planCafeService.getCafeDtosByPlanIdAndMatchType(planId, PlanCafeMatchType.SIMILAR);
-
-        if(findPlan.getMatchType().equals(MatchType.MATCH)) {
-            List<CafeDto> findMatchCafes = planCafeService.getCafeDtosByPlanIdAndMatchType(planId, PlanCafeMatchType.MATCH);
-            return MyPagePlanDetailResponse.of(findPlan, categoryKeywords, findSimilarCafes, findMatchCafes);
-        }
-
-        return MyPagePlanDetailResponse.of(findPlan, categoryKeywords, findSimilarCafes);
-    }
-}
