+++ src/main/java/com/sideproject/hororok/member/application/MyPageService.java
@@ -3,15 +3,23 @@ package com.sideproject.hororok.member.application;
 import com.sideproject.hororok.auth.dto.LoginMember;
 import com.sideproject.hororok.bookmark.application.BookmarkFolderService;
 import com.sideproject.hororok.bookmark.dto.BookmarkFolderDto;
+import com.sideproject.hororok.cafe.dto.CafeDto;
+import com.sideproject.hororok.keword.domain.Keyword;
 import com.sideproject.hororok.keword.domain.enums.Category;
+import com.sideproject.hororok.keword.dto.CategoryKeywordsDto;
 import com.sideproject.hororok.keword.dto.KeywordDto;
 import com.sideproject.hororok.member.domain.Member;
 import com.sideproject.hororok.member.domain.repository.MemberRepository;
 import com.sideproject.hororok.member.dto.MyPagePlanDto;
+import com.sideproject.hororok.member.dto.response.MyPagePlanDetailResponse;
 import com.sideproject.hororok.member.dto.response.MyPagePlanResponse;
 import com.sideproject.hororok.member.dto.response.MyPageProfileResponse;
 import com.sideproject.hororok.member.dto.response.MyPageTagSaveResponse;
+import com.sideproject.hororok.plan.application.PlanCafeService;
+import com.sideproject.hororok.plan.application.PlanKeywordService;
 import com.sideproject.hororok.plan.domain.Plan;
+import com.sideproject.hororok.plan.domain.enums.MatchType;
+import com.sideproject.hororok.plan.domain.enums.PlanCafeMatchType;
 import com.sideproject.hororok.plan.domain.enums.PlanSortBy;
 import com.sideproject.hororok.plan.domain.enums.PlanStatus;
 import com.sideproject.hororok.plan.domain.repository.PlanKeywordRepository;
@@ -41,6 +49,8 @@ public class MyPageService {
     private final ReviewRepository reviewRepository;
     private final PlanKeywordRepository planKeywordRepository;
 
+    private final PlanCafeService planCafeService;
+    private final PlanKeywordService planKeywordService;
     private final BookmarkFolderService bookmarkFolderService;
 
     private static final Integer PLAN_MAX_CNT = 4;
@@ -85,6 +95,21 @@ public class MyPageService {
         }
 
         return new MyPagePlanResponse(plans);
+    }
+
+    public MyPagePlanDetailResponse planDetail(Long planId) {
+
+        Plan findPlan = planRepository.getById(planId);
+        List<Keyword> findKeywords = planKeywordService.getKeywordsByPlanId(planId);
+        CategoryKeywordsDto categoryKeywords = new CategoryKeywordsDto(findKeywords);
+        List<CafeDto> findSimilarCafes = planCafeService.getCafeDtosByPlanIdAndMatchType(planId, PlanCafeMatchType.SIMILAR);
+
+        if(findPlan.getMatchType().equals(MatchType.MATCH)) {
+            List<CafeDto> findMatchCafes = planCafeService.getCafeDtosByPlanIdAndMatchType(planId, PlanCafeMatchType.MATCH);
+            return MyPagePlanDetailResponse.of(findPlan, categoryKeywords, findSimilarCafes, findMatchCafes);
+        }
+
+        return MyPagePlanDetailResponse.of(findPlan, categoryKeywords, findSimilarCafes);
     }
 
     private List<MyPagePlanDto> getPlanByRecent(final LoginMember loginMember,
@@ -137,5 +162,7 @@ public class MyPageService {
 
         return plans;
     }
+
+
 
 }
