+++ src/main/resources/static/docs/api-doc-bookmark-folder.html
@@ -1237,7 +1237,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 303
+Content-Length: 305
 
 {
   "folderId" : 1,
@@ -1248,8 +1248,8 @@ Content-Length: 303
     "cafeId" : 1,
     "cafeName" : "카페 이름",
     "roadAddress" : "OO시 OO구 OO동",
-    "latitude" : 7.541524649806901,
-    "longitude" : 84.54859968364028
+    "latitude" : 129.85896860640403,
+    "longitude" : 12.814107640303016
   } ]
 }</code></pre>
 </div>
