+++ src/main/resources/static/docs/api-doc-myPage.html
@@ -986,7 +986,7 @@ Content-Length: 223
 <div class="sect3">
 <h4 id="_path_parameters">Path parameters</h4>
 <table class="tableblock frame-all grid-all stretch">
-<caption class="title">Table 1. /api/member/myPage/plan/{planId}</caption>
+<caption class="title">Table 1. /api/myPage/plan/{planId}</caption>
 <colgroup>
 <col style="width: 50%;">
 <col style="width: 50%;">
@@ -1009,7 +1009,7 @@ Content-Length: 223
 <h4 id="_http_request_5">Http request</h4>
 <div class="listingblock">
 <div class="content">
-<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /api/member/myPage/plan/1 HTTP/1.1
+<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /api/myPage/plan/1 HTTP/1.1
 Content-Type: application/json
 Authorization: Bearer fake-token
 Accept: application/json
@@ -1026,7 +1026,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 621
+Content-Length: 623
 
 {
   "planId" : 1,
@@ -1047,8 +1047,8 @@ Content-Length: 621
     "name" : "카페 이름",
     "phoneNumber" : "010-1234-5678",
     "roadAddress" : "OO시 OO구 OO동",
-    "longitude" : 84.54859968364028,
-    "latitude" : 7.541524649806901,
+    "longitude" : 12.814107640303016,
+    "latitude" : 129.85896860640403,
     "starRating" : 0,
     "reviewCount" : 0,
     "image" : "//카페이미지"
@@ -1197,7 +1197,7 @@ Content-Length: 621
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2024-04-27 05:39:47 +0900
+Last updated 2024-04-27 05:48:31 +0900
 </div>
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.18.3/highlight.min.js"></script>
