+++ src/main/resources/static/docs/api-doc-plan.html
@@ -475,13 +475,13 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">POST /api/plan HTTP/1.1
 Content-Type: application/json
 Accept: application/json
-Content-Length: 244
+Content-Length: 243
 Host: localhost:8080
 
 {
   "locationName" : "위치 이름",
-  "latitude" : 55.72128878087476,
-  "longitude" : 31.798123539015577,
+  "latitude" : 56.09167874047757,
+  "longitude" : 160.8483214284083,
   "minutes" : 10,
   "date" : "2024-04-27",
   "startTime" : "09:00:00",
@@ -559,7 +559,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 648
+Content-Length: 650
 
 {
   "planId" : 1,
@@ -579,8 +579,8 @@ Content-Length: 648
     "name" : "카페 이름",
     "phoneNumber" : "010-1234-5678",
     "roadAddress" : "OO시 OO구 OO동",
-    "longitude" : 84.54859968364028,
-    "latitude" : 7.541524649806901,
+    "longitude" : 12.814107640303016,
+    "latitude" : 129.85896860640403,
     "starRating" : 0,
     "reviewCount" : 0,
     "image" : "//카페이미지"
